Move constants and log modules to common package
