+++ pyproject.toml
@@ -16,13 +16,13 @@ dependencies = [
     "uvicorn (>=0.34.2,<0.35.0)",
 ]
 
-[project.scripts]
-vigilant = "vigilant.run:main"
-
 [tool.poetry.group.dev.dependencies]
 ipykernel = "^6.29.5"
 pytest = "^8.3.5"
 pytest-cov = "^6.1.1"
+
+[project.scripts]
+vigilant = "vigilant.run:main"
 
 [tool.pytest.ini_options]
 addopts = "--cov=vigilant --cov-branch --cov-report term-missing"
