+++ tests/common/test_constants.py
@@ -1,7 +1,7 @@
 import os
 from typing import Final
 
-from vigilant.constants import MetaSecrets
+from vigilant.common.constants import MetaSecrets
 
 def test_secrets_metaclass() -> None:
     os.environ["ENV_TEST"] = "test_value"
