+++ tests/test_data_collector.py
@@ -8,7 +8,7 @@ from selenium.common.exceptions import WebDriverException
 from selenium.webdriver import ChromeOptions
 
 from vigilant import data_collector
-from vigilant.constants import IOResources
+from vigilant.common.constants import IOResources
 from vigilant.common.exceptions import DriverException, DownloadTimeout
 
 
@@ -60,7 +60,7 @@ def test_build_driver_options(monkeypatch: pytest.MonkeyPatch, tmp_path: Path) -
         "--disable-software-rasterizer",
     ]
 
-    monkeypatch.setattr("vigilant.constants.IOResources.DATA_PATH", tmp_path)
+    monkeypatch.setattr("vigilant.common.constants.IOResources.DATA_PATH", tmp_path)
 
     options = data_collector.build_driver_options()
     preferences = options.experimental_options["prefs"].items()
@@ -109,7 +109,7 @@ def test_clear_resources(monkeypatch: pytest.MonkeyPatch, tmp_path: Path) -> Non
     tmp_file: Path = tmp_path / "test_file.txt"
     tmp_file.write_text("Hesitation is defeat!")
 
-    monkeypatch.setattr("vigilant.constants.IOResources.DATA_PATH", tmp_path)
+    monkeypatch.setattr("vigilant.common.constants.IOResources.DATA_PATH", tmp_path)
 
     data_collector.clear_resources()
 
@@ -145,7 +145,7 @@ def test_get_current_amount(tmp_path: Path, monkeypatch: pytest.MonkeyPatch, moc
 
     mock_driver.find_elements.return_value = found_elements
 
-    monkeypatch.setattr("vigilant.constants.IOResources.DATA_PATH", tmp_path)
+    monkeypatch.setattr("vigilant.common.constants.IOResources.DATA_PATH", tmp_path)
 
     data_collector.get_current_amount(mock_driver)
     amount: str = (IOResources.DATA_PATH / IOResources.AMOUNT_FILENAME).read_text()
@@ -156,7 +156,7 @@ def test_get_current_amount(tmp_path: Path, monkeypatch: pytest.MonkeyPatch, moc
 
 
 def test_get_credit_transactions(tmp_path: Path, monkeypatch: pytest.MonkeyPatch, mock_driver: mock.MagicMock) -> None:
-    monkeypatch.setattr("vigilant.constants.IOResources.DATA_PATH", tmp_path)
+    monkeypatch.setattr("vigilant.common.constants.IOResources.DATA_PATH", tmp_path)
     (tmp_path / "file.xls").write_text("Heasitation is defeat!")
 
     data_collector.get_credit_transactions(mock_driver)
