+++ vigilant/__init__.py
@@ -1,4 +1,4 @@
-from vigilant.log import build_logger
+from vigilant.common.log import build_logger
 
 
 logger = build_logger()
