+++ vigilant/common/constants.py
